git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11262 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/descent.h
@@ -1592,7 +1592,7 @@ class CVisibilityData {
 		ubyte BumpVisitedFlag (void);
 		ubyte BumpProcessedFlag (void);
 		ubyte BumpVisibleFlag (void);
-		int SegmentMayBeVisible (short nStartSeg, short nRadius, int nMaxDist);
+		int SegmentMayBeVisible (short nStartSeg, int nRadius, int nMaxDist);
 		void BuildSegList (CTransformation& transformation, short nStartSeg, int nWindow, bool bIgnoreDoors = false);
 		void InitZRef (int i, int j, int nThread);
 		void QSortZRef (short left, short right);
